refactor: hooks 폴더하위에 common 폴더 생성
